refactor(tests): move the synchronous shutdown helper out of production
RuntimeCoordinator.shutdownAndWait existed only for test defer blocks;
production paths await shutdown() directly. It now lives as an extension
in XcodeMCPTestSupport, leaving the EmbeddedEventLoop compatibility path
in LocalMCPResponder as the only DispatchSemaphore in Sources.
