+++ Tests/XcodeMCPTestSupport/AsyncTestSupport.swift
@@ -1,3 +1,4 @@
+import ProxySession
 import Dispatch
 import Foundation
 import NIO
@@ -191,6 +192,19 @@ package func shutdown(_ group: EventLoopGroup) async {
         group.shutdownGracefully { _ in
             continuation.resume()
         }
+    }
+}
+
+extension RuntimeCoordinator {
+    /// Test-only synchronous teardown for defer blocks; production code
+    /// awaits shutdown() directly.
+    package func shutdownAndWait() {
+        let semaphore = DispatchSemaphore(value: 0)
+        Task.detached(priority: .userInitiated) { [self] in
+            await shutdown()
+            semaphore.signal()
+        }
+        semaphore.wait()
     }
 }
